coeffs: irrationals added to coefficients not the variables
We're using general polynomial routines in SymPy.  Before going to SymPy,
call `symvar` to get a list of variables.  By passing that when creating
the Poly, we explicitly specify the variables (indeterminates) and thus
avoid Poly searching for the domain (a field) for the polynomial.  Thus
we assume polynomial is over the complex numbers.  Fixes #719.